requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
alpha-vantage==2.3.1
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==1.0.0a4
click==8.1.2
colorama==0.4.4
coverage==6.3.2
crashtest==0.3.1
cryptography==36.0.2
cycler==0.11.0
dash==2.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
email-validator==1.2.0
Faker==13.3.4
fastapi==0.75.2
finnhub-python==2.4.13
Flask==2.1.1
Flask-Compress==1.12
fonttools==4.33.3
frozenlist==1.3.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mongoengine==0.24.1
multidict==6.0.2
multitasking==0.0.10
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nose2==0.11.0
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
orjson==3.6.8
packaging==21.3
paho-mqtt==1.6.1
pandas==1.4.2
Pillow==9.0.1
plotly==5.8.0
pptree==3.1
protobuf==3.19.4
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pylev==1.4.0
pymongo==4.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
python-ulid==1.1.0
python3-openid==3.2.0
pytz==2022.1
pywin32==304
PyYAML==6.0
redis==4.2.2
redis-om==0.0.26
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-app-django-mongoengine==1.0.0
social-auth-core==4.2.0
social-auth-storage-mongoengine==1.0.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
starlette==0.17.1
tenacity==8.0.1
threadpoolctl==3.1.0
toml==0.10.2
tradingview-ta==3.2.10
types-redis==4.2.0
types-six==1.16.15
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
webencodings==0.5.1
websocket==0.2.1
websocket-client==1.3.2
websockets==10.3
Werkzeug==2.1.1
whitenoise==6.0.0
wrapt==1.14.0
yarl==1.7.2
yfinance==0.1.60
zope.event==4.5.0
zope.interface==5.4.0
